+++ src/ducks/modules/protocols/create.js
@@ -1,0 +1,24 @@
+import { createProtocol } from '../../../other/protocols';
+import history from '../../../history';
+
+const createProtocolAction = () =>
+  () =>
+    createProtocol()
+      .then(
+        ({ id }) => {
+          history.push(`/edit/${encodeURIComponent(id)}`);
+        },
+      );
+
+const actionCreators = {
+  createProtocol: createProtocolAction,
+};
+
+const actionTypes = {
+};
+
+export {
+  actionCreators,
+  actionTypes,
+};
+
