+++ src/other/protocols/index.js
@@ -1,5 +1,5 @@
 export { default as saveProtocol } from './saveProtocol';
-export { default as openProtocol } from './openProtocol';
+export { default as importProtocol } from './importProtocol';
 export { default as createProtocol } from './createProtocol';
 export { default as loadProtocolData } from './loadProtocolData';
 export { default as importAssetToProtocol } from './importAssetToProtocol';
