--- src/other/protocols/openProtocol.js
@@ -1,22 +1,0 @@
-import path from 'path';
-import { extract } from './archive';
-
-const openProtocol = fileName =>
-  new Promise((resolve) => {
-    if (path.extname(fileName) === '.netcanvas') {
-      extract(fileName)
-        .then((workingPath) => {
-          resolve({
-            archivePath: fileName,
-            workingPath,
-          });
-        });
-    } else {
-      resolve({
-        workingPath: path.dirname(fileName),
-        advanced: true,
-      });
-    }
-  });
-
-export default openProtocol;
